requirements.txt
PyYAML==6.0.2
questionary==2.1.0
prompt_toolkit==3.0.51
